requirements.txt
streamlit==1.28.1  
openai==0.28.0    
python-dotenv==1.0.0  
spacy==3.1.2       
requests==2.31.0   
langchain==0.0.333
PyPDF2==3.0.1